Completed subdomain configuration.  Counselors can only log into their school's domain.
Switched authentication from login field to email field, as this prevents any issue with of making login names local to subdomains, as it is no longer used.
Removed the login field from the counselors table.
Used ApplicationController to apply the require_counselor filter to all controllers.
Added skip_before_filter require_counselor to CounselorSessionsController, since you can't be logged in if you want to log in.